Move the tile renderer
